backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.2
pygments
djangorestframework-jwt
django-cors-headers
django-extensions